Canonicalize test units with @safetestset
Wrap each independent test unit in its own `@safetestset` so it runs in
a fresh module (isolation between tests + world-age safety), matching
the canonical OrdinaryDiffEq structure.

Core (runtests.jl else-branch):
  - The sampler distribution loops + their `test_samples` helper moved
    into a self-contained test/sampler_tests.jl, wrapped in
    `@safetestset "Sampler distribution tests"`.
  - `@testset "BigFloat support"` -> test/bigfloat_tests.jl.
  - `@testset "PassthroughRNG dispatch"` -> test/passthrough_rng_tests.jl.
  - `@testset "Allocation Tests"` keeps including alloc_tests.jl, now
    via `@safetestset`.
Each extracted file carries its own `using`/`import` lines.

QA (test/qa/qa.jl): the Aqua / JET / ExplicitImports plain `@testset`
blocks become `@safetestset`s, each including a self-contained file
(qa_aqua.jl / qa_jet.jl / qa_explicitimports.jl). The include() form is
used so `using JET` runs before the qualified `JET.@test_opt`/`@test_call`
macros are reached.

The GROUP dispatch ladder and all assertions are unchanged. Adds
SafeTestsets to the main test deps and to the QA env (test/qa/Project.toml,
which is activated separately).
