+++ Project.toml
@@ -17,6 +17,7 @@ LogExpFunctions = "0.3.14, 1.0"
 Pkg = "1"
 PrecompileTools = "1.2.0"
 Random = "1.10"
+SafeTestsets = "0.1, 1"
 Statistics = "1"
 Test = "1"
 julia = "1.10"
@@ -28,8 +29,9 @@ Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 ExplicitImports = "7d51a73a-1435-4ff3-83d9-f097790105c7"
 JET = "c3a54625-cd67-489e-a8e7-0a5a0ff4e31b"
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
+SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["AllocCheck", "Aqua", "Statistics", "Test", "Distributions", "JET", "ExplicitImports", "Pkg"]
+test = ["AllocCheck", "Aqua", "Statistics", "Test", "Distributions", "JET", "ExplicitImports", "Pkg", "SafeTestsets"]
