+++ test/qa/Project.toml
@@ -4,6 +4,7 @@ ExplicitImports = "7d51a73a-1435-4ff3-83d9-f097790105c7"
 JET = "c3a54625-cd67-489e-a8e7-0a5a0ff4e31b"
 PoissonRandom = "e409e4f3-bfea-5376-8464-e040bb5c01ab"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [sources]
@@ -14,5 +15,6 @@ Aqua = "0.8"
 ExplicitImports = "1.14.0"
 JET = "0.9, 0.10, 0.11"
 Random = "1.10"
+SafeTestsets = "0.1, 1"
 Test = "1"
 julia = "1.10"
