+++ test/qa/qa.jl
@@ -1,36 +1,13 @@
-using PoissonRandom, Aqua, JET, ExplicitImports
-using Random
-using Test
+using SafeTestsets
 
-@testset "Aqua" begin
-    Aqua.find_persistent_tasks_deps(PoissonRandom)
-    Aqua.test_ambiguities(PoissonRandom, recursive = false)
-    Aqua.test_deps_compat(PoissonRandom)
-    Aqua.test_piracies(
-        PoissonRandom,
-        treat_as_own = []
-    )
-    Aqua.test_project_extras(PoissonRandom)
-    Aqua.test_stale_deps(PoissonRandom)
-    Aqua.test_unbound_args(PoissonRandom)
-    Aqua.test_undefined_exports(PoissonRandom)
+@safetestset "Aqua" begin
+    include("qa_aqua.jl")
 end
 
-@testset "JET static analysis" begin
-    @testset "Type stability" begin
-        JET.@test_opt target_modules = (PoissonRandom,) pois_rand(10.0)
-        JET.@test_opt target_modules = (PoissonRandom,) pois_rand(Random.default_rng(), 10.0)
-        JET.@test_opt target_modules = (PoissonRandom,) pois_rand(PassthroughRNG(), 10.0)
-    end
-
-    @testset "Error analysis" begin
-        JET.@test_call target_modules = (PoissonRandom,) pois_rand(10.0)
-        JET.@test_call target_modules = (PoissonRandom,) pois_rand(Random.default_rng(), 10.0)
-        JET.@test_call target_modules = (PoissonRandom,) pois_rand(PassthroughRNG(), 10.0)
-    end
+@safetestset "JET static analysis" begin
+    include("qa_jet.jl")
 end
 
-@testset "ExplicitImports" begin
-    @test check_no_implicit_imports(PoissonRandom) === nothing
-    @test check_no_stale_explicit_imports(PoissonRandom) === nothing
+@safetestset "ExplicitImports" begin
+    include("qa_explicitimports.jl")
 end
