+++ test/qa/qa_explicitimports.jl
@@ -1,0 +1,5 @@
+using PoissonRandom, ExplicitImports
+using Test
+
+@test check_no_implicit_imports(PoissonRandom) === nothing
+@test check_no_stale_explicit_imports(PoissonRandom) === nothing
